+++ hiera5/data_source_hiera_array_test.go
@@ -28,7 +28,7 @@ func TestAccDataSourceHiera5Array_Basic(t *testing.T) {
 				Check: resource.ComposeTestCheckFunc(
 					testAccDataSourceHiera5ArrayCheck(keyUnavailable),
 				),
-				ExpectError: regexp.MustCompile("unexpected end of JSON input"),
+				ExpectError: regexp.MustCompile("Key '" + keyUnavailable + "' not found"),
 			},
 		},
 	})