+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 ## 0.1.1 (Unrelased)
+
+BUG FIXES:
+
+* Ensure that `lookup` always returns a valid JSON or an error
+
 ## 0.1.0 (February 11, 2020)
 
 FEATURES: